set ngl_viewer in a seperate TS file
